+++ WebApplicationRefregitzTow/DrawSoldire.cs
@@ -62,7 +62,7 @@ namespace RefrigtzW
         public bool MaxFound(ref bool MaxNotFound)
         {
             //long Time = TimeElapced.TimeNow();Spaces++;
-            double a = ReturnHuristic();
+            int a = ReturnHuristic();
             if (MaxHuristicxS < a)
             {
                 Object O2 = new Object();
@@ -81,11 +81,11 @@ namespace RefrigtzW
             ////{ AllDraw.OutPut.Append("\r\n");for (int l = 0; l < Spaces; l++) AllDraw.OutPut.Append(Space);  AllDraw.OutPut.Append("MaxFound:" + (TimeElapced.TimeNow() - Time).ToString());}Spaces--;
             return false;
         }
-        public double ReturnHuristic()
+        public int ReturnHuristic()
         {
             int HaveKilled = 0;
             //long Time = TimeElapced.TimeNow();Spaces++;
-            double a = 0;
+            int a = 0;
             for (var ii = 0; ii < AllDraw.SodierMovments; ii++)
 
                 a += SoldierThinking[ii].ReturnHuristic(-1, -1, Order, false,ref HaveKilled);
